Remove V2 GrpcTitusServiceGateway implementation and the router
In this change `V3GrpcTitusServiceGateway` is not removed yet, as it contains
non-trivial logic that must be moved to other classes. This requires at least:
* refactoring V3JobOperations API to support the cursor based pagination mechanism
* refactoring JobSubmitLimiter, and moving the unique job cluster id constraint into the V3JobOperations service
