--- titus-server-master/src/main/java/com/netflix/titus/master/RuntimeLimitValidator.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2018 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.titus.master;
-
-import com.netflix.titus.api.endpoint.v2.rest.representation.TitusJobType;
-import com.netflix.titus.api.jobmanager.model.job.sanitizer.JobConfiguration;
-import com.netflix.titus.master.endpoint.v2.rest.representation.TitusJobSpec;
-import com.netflix.titus.master.endpoint.v2.validator.TitusJobSpecValidators;
-
-public class RuntimeLimitValidator implements TitusJobSpecValidators.Validator {
-    private final JobConfiguration jobConfiguration;
-
-    public RuntimeLimitValidator(JobConfiguration jobConfiguration) {
-        this.jobConfiguration = jobConfiguration;
-    }
-
-    @Override
-    public boolean isValid(TitusJobSpec titusJobSpec) {
-        return titusJobSpec.getType() == TitusJobType.service ||
-                (titusJobSpec.getType() == TitusJobType.batch &&
-                        titusJobSpec.getRuntimeLimitSecs() != null &&
-                        titusJobSpec.getRuntimeLimitSecs() >= 0 &&
-                        titusJobSpec.getRuntimeLimitSecs() <= jobConfiguration.getMaxRuntimeLimitSec());
-    }
-}
