--- titus-server-master/src/main/java/com/netflix/titus/master/endpoint/v2/validator/ValidatorConfiguration.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2018 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.titus.master.endpoint.v2.validator;
-
-import java.util.List;
-
-import com.netflix.archaius.api.annotations.Configuration;
-import com.netflix.archaius.api.annotations.PropertyName;
-
-import static java.util.Arrays.asList;
-
-@Configuration(prefix = "titusMaster.endpoint.v2.validator")
-public interface ValidatorConfiguration {
-
-    List<String> DEFAULT_INVALID_EFS_MOUNT_POINTS = asList("/data", "/logs");
-
-    @PropertyName(name = "efs.invalidMountPoints")
-    default List<String> getInvalidEfsMountPoints() {
-        return DEFAULT_INVALID_EFS_MOUNT_POINTS;
-    }
-}
