+++ titus-server-master/src/main/java/com/netflix/titus/master/jobmanager/endpoint/v3/V3EndpointModule.java
@@ -22,31 +22,15 @@ import javax.inject.Singleton;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.TypeLiteral;
-import com.netflix.titus.api.agent.service.AgentManagementService;
 import com.netflix.titus.api.jobmanager.model.job.Task;
 import com.netflix.titus.api.jobmanager.service.V3JobOperations;
 import com.netflix.titus.common.model.sanitizer.EntitySanitizer;
 import com.netflix.titus.common.runtime.TitusRuntime;
-import com.netflix.titus.grpc.protogen.Job;
-import com.netflix.titus.grpc.protogen.JobDescriptor;
-import com.netflix.titus.grpc.protogen.JobDescriptor.JobSpecCase;
 import com.netflix.titus.grpc.protogen.JobManagementServiceGrpc.JobManagementServiceImplBase;
-import com.netflix.titus.grpc.protogen.TaskStatus;
-import com.netflix.titus.master.ApiOperations;
-import com.netflix.titus.master.cluster.LeaderActivator;
-import com.netflix.titus.master.config.CellInfoResolver;
-import com.netflix.titus.master.config.MasterConfiguration;
-import com.netflix.titus.master.endpoint.TitusServiceGateway;
-import com.netflix.titus.master.endpoint.adapter.LegacyTitusServiceGatewayGuard;
-import com.netflix.titus.master.endpoint.grpc.GrpcEndpointConfiguration;
 import com.netflix.titus.master.jobmanager.endpoint.v3.grpc.DefaultJobManagementServiceGrpc;
 import com.netflix.titus.master.jobmanager.endpoint.v3.grpc.gateway.GrpcTitusServiceGateway;
-import com.netflix.titus.master.jobmanager.endpoint.v3.grpc.gateway.RoutingGrpcTitusServiceGateway;
-import com.netflix.titus.master.jobmanager.endpoint.v3.grpc.gateway.V2GrpcTitusServiceGateway;
 import com.netflix.titus.master.jobmanager.endpoint.v3.grpc.gateway.V3GrpcTitusServiceGateway;
 import com.netflix.titus.master.jobmanager.service.limiter.JobSubmitLimiter;
-import com.netflix.titus.master.master.MasterMonitor;
-import com.netflix.titus.master.service.management.ApplicationSlaManagementService;
 import com.netflix.titus.runtime.endpoint.common.LogStorageInfo;
 import com.netflix.titus.runtime.endpoint.metadata.CallMetadataResolver;
 import com.netflix.titus.runtime.endpoint.metadata.SimpleGrpcCallMetadataResolver;
@@ -61,39 +45,17 @@ public class V3EndpointModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(GrpcTitusServiceGateway.class).to(V2GrpcTitusServiceGateway.class);
         bind(CallMetadataResolver.class).to(SimpleGrpcCallMetadataResolver.class);
         bind(JobManagementServiceImplBase.class).to(DefaultJobManagementServiceGrpc.class);
     }
 
     @Provides
     @Singleton
-    @Named(RoutingGrpcTitusServiceGateway.NAME_V3_ENGINE_GATEWAY)
     public GrpcTitusServiceGateway getV3ServiceGateway(V3JobOperations jobOperations,
                                                        JobSubmitLimiter jobSubmitLimiter,
                                                        LogStorageInfo<Task> v3LogStorage,
                                                        @Named(JOB_STRICT_SANITIZER) EntitySanitizer entitySanitizer,
                                                        TitusRuntime titusRuntime) {
         return new V3GrpcTitusServiceGateway(jobOperations, jobSubmitLimiter, v3LogStorage, entitySanitizer, titusRuntime);
-    }
-
-    @Provides
-    @Singleton
-    public TitusServiceGateway<String, JobDescriptor, JobSpecCase, Job, com.netflix.titus.grpc.protogen.Task, TaskStatus.TaskState> getV3RoutingServiceGateway(
-            @Named(RoutingGrpcTitusServiceGateway.NAME_V2_ENGINE_GATEWAY) GrpcTitusServiceGateway v2EngineGateway,
-            @Named(RoutingGrpcTitusServiceGateway.NAME_V3_ENGINE_GATEWAY) GrpcTitusServiceGateway v3EngineGateway,
-            AgentManagementService agentManagementService,
-            ApplicationSlaManagementService capacityGroupManagement,
-            ApiOperations apiOperations,
-            MasterMonitor masterMonitor,
-            LeaderActivator leaderActivator,
-            GrpcEndpointConfiguration configuration,
-            MasterConfiguration masterConfiguration,
-            CellInfoResolver cellInfoResolver) {
-        RoutingGrpcTitusServiceGateway serviceGateway = new RoutingGrpcTitusServiceGateway(
-                v2EngineGateway, v3EngineGateway, agentManagementService, capacityGroupManagement,
-                configuration, masterConfiguration, cellInfoResolver
-        );
-        return new LegacyTitusServiceGatewayGuard<>(serviceGateway, apiOperations, masterMonitor, leaderActivator);
     }
 }
