+++ titus-server-master/src/main/java/com/netflix/titus/master/jobmanager/endpoint/v3/grpc/DefaultJobManagementServiceGrpc.java
@@ -27,7 +27,12 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 
 import com.google.protobuf.Empty;
+import com.netflix.titus.api.agent.service.AgentManagementService;
+import com.netflix.titus.api.jobmanager.model.job.ContainerResources;
+import com.netflix.titus.api.jobmanager.service.JobManagerException;
 import com.netflix.titus.api.model.Pagination;
+import com.netflix.titus.api.model.ResourceDimension;
+import com.netflix.titus.api.model.Tier;
 import com.netflix.titus.api.service.TitusServiceException;
 import com.netflix.titus.common.util.ProtobufCopy;
 import com.netflix.titus.common.util.StringExt;
@@ -51,10 +56,18 @@ import com.netflix.titus.grpc.protogen.TaskKillRequest;
 import com.netflix.titus.grpc.protogen.TaskQuery;
 import com.netflix.titus.grpc.protogen.TaskQueryResult;
 import com.netflix.titus.grpc.protogen.TaskStatus;
+import com.netflix.titus.master.config.CellInfoResolver;
 import com.netflix.titus.master.endpoint.TitusServiceGateway;
+import com.netflix.titus.master.endpoint.common.CellDecorator;
+import com.netflix.titus.master.endpoint.grpc.GrpcEndpointConfiguration;
+import com.netflix.titus.master.jobmanager.endpoint.v3.grpc.gateway.GrpcTitusServiceGateway;
+import com.netflix.titus.master.jobmanager.service.JobManagerUtil;
+import com.netflix.titus.master.model.ResourceDimensions;
+import com.netflix.titus.master.service.management.ApplicationSlaManagementService;
 import com.netflix.titus.runtime.endpoint.JobQueryCriteria;
 import com.netflix.titus.runtime.endpoint.metadata.CallMetadata;
 import com.netflix.titus.runtime.endpoint.metadata.CallMetadataResolver;
+import com.netflix.titus.runtime.endpoint.v3.grpc.V3GrpcModelConverters;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
 import io.grpc.stub.ServerCallStreamObserver;
@@ -64,39 +77,67 @@ import org.slf4j.LoggerFactory;
 import rx.Observable;
 import rx.Subscription;
 
+import static com.netflix.titus.runtime.connector.jobmanager.JobManagementClient.JOB_MINIMUM_FIELD_SET;
+import static com.netflix.titus.runtime.connector.jobmanager.JobManagementClient.TASK_MINIMUM_FIELD_SET;
 import static com.netflix.titus.runtime.endpoint.common.grpc.CommonGrpcModelConverters.toGrpcPagination;
 import static com.netflix.titus.runtime.endpoint.common.grpc.CommonGrpcModelConverters.toJobQueryCriteria;
 import static com.netflix.titus.runtime.endpoint.common.grpc.CommonGrpcModelConverters.toPage;
 import static com.netflix.titus.runtime.endpoint.common.grpc.GrpcUtil.safeOnError;
 import static com.netflix.titus.runtime.endpoint.v3.grpc.TitusPaginationUtils.checkPageIsValid;
-import static com.netflix.titus.runtime.connector.jobmanager.JobManagementClient.JOB_MINIMUM_FIELD_SET;
-import static com.netflix.titus.runtime.connector.jobmanager.JobManagementClient.TASK_MINIMUM_FIELD_SET;
 
 @Singleton
 public class DefaultJobManagementServiceGrpc extends JobManagementServiceGrpc.JobManagementServiceImplBase {
 
     private static final Logger logger = LoggerFactory.getLogger(DefaultJobManagementServiceGrpc.class);
 
+    private final GrpcEndpointConfiguration configuration;
+    private final AgentManagementService agentManagementService;
+    private final ApplicationSlaManagementService capacityGroupService;
     private final TitusServiceGateway<String, JobDescriptor, JobSpecCase, Job, Task, TaskStatus.TaskState> serviceGateway;
     private final CallMetadataResolver callMetadataResolver;
+    private final CellDecorator cellDecorator;
 
     @Inject
-    public DefaultJobManagementServiceGrpc(TitusServiceGateway<String, JobDescriptor, JobSpecCase, Job, Task, TaskStatus.TaskState> serviceGateway,
-                                           CallMetadataResolver callMetadataResolver) {
+    public DefaultJobManagementServiceGrpc(GrpcEndpointConfiguration configuration,
+                                           AgentManagementService agentManagementService,
+                                           ApplicationSlaManagementService capacityGroupService,
+                                           GrpcTitusServiceGateway serviceGateway,
+                                           CallMetadataResolver callMetadataResolver,
+                                           CellInfoResolver cellInfoResolver) {
+        this.configuration = configuration;
+        this.agentManagementService = agentManagementService;
+        this.capacityGroupService = capacityGroupService;
         this.serviceGateway = serviceGateway;
         this.callMetadataResolver = callMetadataResolver;
+        this.cellDecorator = new CellDecorator(cellInfoResolver::getCellName);
     }
 
     @Override
-    public void createJob(JobDescriptor request, StreamObserver<JobId> responseObserver) {
-        execute(responseObserver, callMetadata ->
-                serviceGateway.createJob(
-                        request
-                ).subscribe(
-                        jobId -> responseObserver.onNext(JobId.newBuilder().setId(jobId).build()),
-                        e -> safeOnError(logger, e, responseObserver),
-                        responseObserver::onCompleted
-                ));
+    public void createJob(JobDescriptor jobDescriptor, StreamObserver<JobId> responseObserver) {
+        execute(responseObserver, callMetadata -> {
+            if (configuration.isJobSizeValidationEnabled()) {
+                com.netflix.titus.api.jobmanager.model.job.JobDescriptor coreJobDescriptor = V3GrpcModelConverters.toCoreJobDescriptor(jobDescriptor);
+
+                Tier tier = findTier(coreJobDescriptor);
+                ResourceDimension requestedResources = toResourceDimension(coreJobDescriptor.getContainer().getContainerResources());
+                List<ResourceDimension> tierResourceLimits = getTierResourceLimits(tier);
+                if (isTooLarge(requestedResources, tierResourceLimits)) {
+                    safeOnError(logger,
+                            JobManagerException.invalidContainerResources(tier, requestedResources, tierResourceLimits),
+                            responseObserver
+                    );
+                    return;
+                }
+            }
+
+            JobDescriptor withCellInfo = cellDecorator.ensureCellInfo(jobDescriptor);
+
+            serviceGateway.createJob(withCellInfo).subscribe(
+                    jobId -> responseObserver.onNext(JobId.newBuilder().setId(jobId).build()),
+                    e -> safeOnError(logger, e, responseObserver),
+                    responseObserver::onCompleted
+            );
+        });
     }
 
     @Override
@@ -319,6 +360,16 @@ public class DefaultJobManagementServiceGrpc extends JobManagementServiceGrpc.Jo
                 .build();
     }
 
+    private ResourceDimension toResourceDimension(ContainerResources containerResources) {
+        return ResourceDimension.newBuilder()
+                .withCpus(containerResources.getCpu())
+                .withGpu(containerResources.getGpu())
+                .withMemoryMB(containerResources.getMemoryMB())
+                .withDiskMB(containerResources.getDiskMB())
+                .withNetworkMbs(containerResources.getNetworkMbps())
+                .build();
+    }
+
     private String toReasonString(CallMetadata callMetadata) {
         StringBuilder builder = new StringBuilder();
         builder.append("calledBy=").append(callMetadata.getCallerId());
@@ -341,5 +392,22 @@ public class DefaultJobManagementServiceGrpc extends JobManagementServiceGrpc.Jo
         }
 
         return builder.toString();
+    }
+
+    private Tier findTier(com.netflix.titus.api.jobmanager.model.job.JobDescriptor jobDescriptor) {
+        return JobManagerUtil.getTierAssignment(jobDescriptor, capacityGroupService).getLeft();
+    }
+
+    private List<ResourceDimension> getTierResourceLimits(Tier tier) {
+        return agentManagementService.getInstanceGroups().stream()
+                .filter(instanceGroup -> instanceGroup.getTier().equals(tier))
+                .map(instanceGroup ->
+                        agentManagementService.findResourceLimits(instanceGroup.getInstanceType()).orElse(instanceGroup.getResourceDimension())
+                )
+                .collect(Collectors.toList());
+    }
+
+    private boolean isTooLarge(ResourceDimension requestedResources, List<ResourceDimension> tierResourceLimits) {
+        return tierResourceLimits.stream().noneMatch(limit -> ResourceDimensions.isBigger(limit, requestedResources));
     }
 }
